+++ src/gulpfile.js
@@ -300,7 +300,7 @@ gulp.task('pub', ['pack_build'], function () {
     inquirer.prompt(questions).then(function (answers) {
       var pkg = util.getPkg();
       pkg.version = answers.version;
-      file.writeFileFromString(JSON.stringify(pkg, null, ' '), 'package.json');
+      file.writeFileFromString(JSON.stringify(pkg, null, '  '), 'package.json');
       console.log(colors.info('#### Git Info ####'));
       spawn.sync('git', ['add', '.'], { stdio: 'inherit' });
       spawn.sync('git', ['commit', '-m', 'ver. ' + pkg.version], { stdio: 'inherit' });